Convert bash tests to rust integration tests (#333)
* tests: convert bash integration tests to rs

* lint: clippy and formatting issues

* fix: to past tense


* fix: add traces for errors in tests

* fix: minor fixes on test and basic api

* add: revert tx block

---------
