+++ crates/bin/prove_block/Cargo.toml
@@ -5,6 +5,7 @@ edition.workspace = true
 repository.workspace = true
 license-file.workspace = true
 
+
 [dependencies]
 anyhow = { workspace = true }
 blockifier = { workspace = true }
@@ -30,3 +31,6 @@ starknet-os-types = { workspace = true }
 starknet-types-core = { workspace = true }
 tokio = { workspace = true }
 num-bigint = { workspace = true }
+
+[dev-dependencies]
+rstest = { workspace = true }
